==> codes/08-filter.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os

# 원본 데이터 경로
original_data_path = 'C:/Users/yeon1/data-noise-research2/output'

# 필터링된 데이터 파일들이 있는 경로
# target_ratio는 anomaly ratio를 뜻함
target_ratio = 'target_ratio_10' # change (1) 
filtered_data_dir = 'C:/Users/yeon1/Leverage_Eventlog/Leverage_Ko/filtered/' + f'{target_ratio}'

# 출력 디렉토리 설정
output_dir = 'C:/Users/yeon1/data-noise-research2/washed/' + f'{target_ratio}'
if not os.path.exists(output_dir):
    os.makedirs(output_dir)

# 필터링된 데이터를 저장할 세트 초기화
cases_to_remove_leverage2 = set()

# 필터링된 데이터 파일들을 순회하며 세트에 추가
filtered_filenames = sorted(os.listdir(filtered_data_dir))
for filtered_filename in os.listdir(filtered_data_dir):
    if filtered_filename.endswith('.csv'):
        filtered_file_path = os.path.join(filtered_data_dir, filtered_filename)
        filtered_data = pd.read_csv(filtered_file_path)
        cases_to_remove_leverage2.update(filtered_data['caseid'])

# 원본 데이터 파일들을 순회하며 필터링된 데이터를 적용
original_filenames = sorted(os.listdir(original_data_path))
for filename in os.listdir(original_data_path):
    if filename.endswith('.csv'):
        file_path = os.path.join(original_data_path, filename)
        
        # 원본 데이터 읽기
        original_data = pd.read_csv(file_path)
        
        # 필터링된 데이터를 원본 데이터에 적용
        filtered_original_data_leverage2 = original_data[original_data['Case'].isin(cases_to_remove_leverage2)]
        
        # 출력 파일 경로 설정
        output_file_path = os.path.join(output_dir, f'washed_{filename}')
        
        # 필터링된 데이터 저장
        filtered_original_data_leverage2.to_csv(output_file_path, index=False)

        print(f'{output_file_path}', "Data filtering and saving completed.")
