==> codes/09-temp.ipynb <==
# temp 1~30까지 생성

file_list = os.listdir(washed_path + '/target_ratio_10') # change ratio
csv_list = [s for s in file_list if '.csv' in s]  # change '.csv'
print(csv_list) # 먼저 확인하기
len(csv_list)

# noisy data set에 prefix length를 적용하는 코드, temp가 완성된다.
for prefix_length in range(1, 31):
    for event_log in csv_list:
        df = pd.read_csv(washed_path + '/target_ratio_10' + '\\' + event_log) # output path
        prefix_df = df.groupby('Case').head(prefix_length)  # prefix
        prefix_df.to_csv(washed_temp + f'\\prefix{prefix_length}_target_ratio_10_{event_log}', index=False)

# temp 1~30까지 생성

file_list = os.listdir(washed_path + '/target_ratio_20') # change ratio
csv_list = [s for s in file_list if '.csv' in s]  # change '.csv'
print(csv_list) # 먼저 확인하기
len(csv_list)

# noisy data set에 prefix length를 적용하는 코드, temp가 완성된다.
for prefix_length in range(1, 31):
    for event_log in csv_list:
        df = pd.read_csv(washed_path + '/target_ratio_20' + '\\' + event_log) # output path
        prefix_df = df.groupby('Case').head(prefix_length)  # prefix
        prefix_df.to_csv(washed_temp + f'\\prefix{prefix_length}_target_ratio_20_{event_log}', index=False)

# temp 1~30까지 생성

file_list = os.listdir(washed_path + '/target_ratio_30') # change ratio
csv_list = [s for s in file_list if '.csv' in s]  # change '.csv'
print(csv_list) # 먼저 확인하기
len(csv_list)

# noisy data set에 prefix length를 적용하는 코드, temp가 완성된다.
for prefix_length in range(1, 31):
    for event_log in csv_list:
        df = pd.read_csv(washed_path + '/target_ratio_30' + '\\' + event_log) # output path
        prefix_df = df.groupby('Case').head(prefix_length)  # prefix
        prefix_df.to_csv(washed_temp + f'\\prefix{prefix_length}_target_ratio_30_{event_log}', index=False)

# temp 1~30까지 생성

file_list = os.listdir(washed_path + '/target_ratio_40') # change ratio
csv_list = [s for s in file_list if '.csv' in s]  # change '.csv'
print(csv_list) # 먼저 확인하기
len(csv_list)

# noisy data set에 prefix length를 적용하는 코드, temp가 완성된다.
for prefix_length in range(1, 31):
    for event_log in csv_list:
        df = pd.read_csv(washed_path + '/target_ratio_40' + '\\' + event_log) # output path
        prefix_df = df.groupby('Case').head(prefix_length)  # prefix
        prefix_df.to_csv(washed_temp + f'\\prefix{prefix_length}_target_ratio_40_{event_log}', index=False)
