==> codes/10-concat.ipynb <==
# clean concat

file_list = os.listdir(washed_temp)
noise_type = ['insert0.003', 'insert0.006', 'insert0.009',
            'moved0.003', 'moved0.006', 'moved0.009',
            'replace0.003', 'replace0.006', 'replace0.009',
            'rework0.003', 'rework0.006', 'rework0.009',
            'skip0.003', 'skip0.006', 'skip0.009']

for i in range(0, len(noise_type)):
    csv_list = [s for s in file_list if 'target_ratio_10' in s and # change (1)
                 noise_type[i] in s and
                 'BPIC15_1' in s and # change (2)
                 s.endswith('.csv')]

    # csv_list를 prefix1부터 prefix30까지 정렬하기
    csv_list = sorted(csv_list, key=lambda x: int(x.split('prefix')[1].split('_')[0]))

    encoder_df = pd.DataFrame()  # 빈 데이터프레임 생성

    for k in range(0, len(csv_list)):
        df = pd.read_csv(washed_temp + '\\' + csv_list[k])

        # Timestamp를 기준으로 앞 쪽 80%를 train 데이터로 선택
        df['Timestamp'] = pd.to_datetime(df['Timestamp'])
        df_sorted = df.sort_values(by='Timestamp')
        train_length = int(len(df_sorted) * 0.8)
        train_data = df_sorted.iloc[:train_length]

        # 인코딩 수행
        encoder = Aggregate(case_id_col=['Case'], cat_cols=['Activity', 'label'], boolean=False)
        encoder_df_temp = encoder.fit_transform(df)
        try: encoder_df_temp['label'] = encoder_df_temp['label_deviant'].apply(lambda x: 0 if x > 0 else 1) # deviant 0, regular 1
        except: encoder_df_temp['label'] = encoder_df_temp['label_regular'].apply(lambda x: 1 if x > 0 else 0)
        try: encoder_df_temp.drop(['label_deviant', 'label_regular'], axis=1, inplace=True)
        except: encoder_df_temp.drop(['label_regular'], axis=1, inplace=True)

        # train 데이터와 test 데이터 분리
        encoder_df_temp['set'] = np.where(encoder_df_temp.index.isin(train_data['Case']), 'train', 'test')
        train = encoder_df_temp[encoder_df_temp['set'] == 'train']
        train = train.drop(['set'], axis=1)

        # train 데이터를 이어붙이기
        encoder_df = pd.concat([encoder_df, train])

        encoder_df = encoder_df.fillna(0)
        encoder_df = encoder_df.astype(int)
    
        # 결과 저장
        encoder_df.to_csv(concat_path + "\\" + noise_type[i] + '_BPIC15_1_' + 'target_ratio_10_concat.csv', index=True) # change (3)

# clean concat

file_list = os.listdir(washed_temp)
noise_type = ['insert0.003', 'insert0.006', 'insert0.009',
            'moved0.003', 'moved0.006', 'moved0.009',
            'replace0.003', 'replace0.006', 'replace0.009',
            'rework0.003', 'rework0.006', 'rework0.009',
            'skip0.003', 'skip0.006', 'skip0.009']

for i in range(0, len(noise_type)):
    csv_list = [s for s in file_list if 'target_ratio_20' in s and # change (1)
                 noise_type[i] in s and
                 'BPIC15_1' in s and # change (2)
                 s.endswith('.csv')]

    # csv_list를 prefix1부터 prefix30까지 정렬하기
    csv_list = sorted(csv_list, key=lambda x: int(x.split('prefix')[1].split('_')[0]))

    encoder_df = pd.DataFrame()  # 빈 데이터프레임 생성

    for k in range(0, len(csv_list)):
        df = pd.read_csv(washed_temp + '\\' + csv_list[k])

        # Timestamp를 기준으로 앞 쪽 80%를 train 데이터로 선택
        df['Timestamp'] = pd.to_datetime(df['Timestamp'])
        df_sorted = df.sort_values(by='Timestamp')
        train_length = int(len(df_sorted) * 0.8)
        train_data = df_sorted.iloc[:train_length]

        # 인코딩 수행
        encoder = Aggregate(case_id_col=['Case'], cat_cols=['Activity', 'label'], boolean=False)
        encoder_df_temp = encoder.fit_transform(df)
        try: encoder_df_temp['label'] = encoder_df_temp['label_deviant'].apply(lambda x: 0 if x > 0 else 1) # deviant 0, regular 1
        except: encoder_df_temp['label'] = encoder_df_temp['label_regular'].apply(lambda x: 1 if x > 0 else 0)
        try: encoder_df_temp.drop(['label_deviant', 'label_regular'], axis=1, inplace=True)
        except: encoder_df_temp.drop(['label_regular'], axis=1, inplace=True)

        # train 데이터와 test 데이터 분리
        encoder_df_temp['set'] = np.where(encoder_df_temp.index.isin(train_data['Case']), 'train', 'test')
        train = encoder_df_temp[encoder_df_temp['set'] == 'train']
        train = train.drop(['set'], axis=1)

        # train 데이터를 이어붙이기
        encoder_df = pd.concat([encoder_df, train])

        encoder_df = encoder_df.fillna(0)
        encoder_df = encoder_df.astype(int)
    
        # 결과 저장
        encoder_df.to_csv(concat_path + "\\" + noise_type[i] + '_BPIC15_1_' + 'target_ratio_20_concat.csv', index=True) # change (3)

# clean concat

file_list = os.listdir(washed_temp)
noise_type = ['insert0.003', 'insert0.006', 'insert0.009',
            'moved0.003', 'moved0.006', 'moved0.009',
            'replace0.003', 'replace0.006', 'replace0.009',
            'rework0.003', 'rework0.006', 'rework0.009',
            'skip0.003', 'skip0.006', 'skip0.009']

for i in range(0, len(noise_type)):
    csv_list = [s for s in file_list if 'target_ratio_30' in s and # change (1)
                 noise_type[i] in s and
                 'BPIC15_1' in s and # change (2)
                 s.endswith('.csv')]

    # csv_list를 prefix1부터 prefix30까지 정렬하기
    csv_list = sorted(csv_list, key=lambda x: int(x.split('prefix')[1].split('_')[0]))

    encoder_df = pd.DataFrame()  # 빈 데이터프레임 생성

    for k in range(0, len(csv_list)):
        df = pd.read_csv(washed_temp + '\\' + csv_list[k])

        # Timestamp를 기준으로 앞 쪽 80%를 train 데이터로 선택
        df['Timestamp'] = pd.to_datetime(df['Timestamp'])
        df_sorted = df.sort_values(by='Timestamp')
        train_length = int(len(df_sorted) * 0.8)
        train_data = df_sorted.iloc[:train_length]

        # 인코딩 수행
        encoder = Aggregate(case_id_col=['Case'], cat_cols=['Activity', 'label'], boolean=False)
        encoder_df_temp = encoder.fit_transform(df)
        try: encoder_df_temp['label'] = encoder_df_temp['label_deviant'].apply(lambda x: 0 if x > 0 else 1) # deviant 0, regular 1
        except: encoder_df_temp['label'] = encoder_df_temp['label_regular'].apply(lambda x: 1 if x > 0 else 0)
        try: encoder_df_temp.drop(['label_deviant', 'label_regular'], axis=1, inplace=True)
        except: encoder_df_temp.drop(['label_regular'], axis=1, inplace=True)

        # train 데이터와 test 데이터 분리
        encoder_df_temp['set'] = np.where(encoder_df_temp.index.isin(train_data['Case']), 'train', 'test')
        train = encoder_df_temp[encoder_df_temp['set'] == 'train']
        train = train.drop(['set'], axis=1)

        # train 데이터를 이어붙이기
        encoder_df = pd.concat([encoder_df, train])

        encoder_df = encoder_df.fillna(0)
        encoder_df = encoder_df.astype(int)
    
        # 결과 저장
        encoder_df.to_csv(concat_path + "\\" + noise_type[i] + '_BPIC15_1_' + 'target_ratio_30_concat.csv', index=True) # change (3)

# clean concat

file_list = os.listdir(washed_temp)
noise_type = ['insert0.003', 'insert0.006', 'insert0.009',
            'moved0.003', 'moved0.006', 'moved0.009',
            'replace0.003', 'replace0.006', 'replace0.009',
            'rework0.003', 'rework0.006', 'rework0.009',
            'skip0.003', 'skip0.006', 'skip0.009']

for i in range(0, len(noise_type)):
    csv_list = [s for s in file_list if 'target_ratio_40' in s and # change (1)
                 noise_type[i] in s and
                 'BPIC15_1' in s and # change (2)
                 s.endswith('.csv')]

    # csv_list를 prefix1부터 prefix30까지 정렬하기
    csv_list = sorted(csv_list, key=lambda x: int(x.split('prefix')[1].split('_')[0]))

    encoder_df = pd.DataFrame()  # 빈 데이터프레임 생성

    for k in range(0, len(csv_list)):
        df = pd.read_csv(washed_temp + '\\' + csv_list[k])

        # Timestamp를 기준으로 앞 쪽 80%를 train 데이터로 선택
        df['Timestamp'] = pd.to_datetime(df['Timestamp'])
        df_sorted = df.sort_values(by='Timestamp')
        train_length = int(len(df_sorted) * 0.8)
        train_data = df_sorted.iloc[:train_length]
        
        # 인코딩 수행
        encoder = Aggregate(case_id_col=['Case'], cat_cols=['Activity', 'label'], boolean=False)
        encoder_df_temp = encoder.fit_transform(df)
        try: encoder_df_temp['label'] = encoder_df_temp['label_deviant'].apply(lambda x: 0 if x > 0 else 1) # deviant 0, regular 1
        except: encoder_df_temp['label'] = encoder_df_temp['label_regular'].apply(lambda x: 1 if x > 0 else 0)
        try: encoder_df_temp.drop(['label_deviant', 'label_regular'], axis=1, inplace=True)
        except: encoder_df_temp.drop(['label_regular'], axis=1, inplace=True)

        # train 데이터와 test 데이터 분리
        encoder_df_temp['set'] = np.where(encoder_df_temp.index.isin(train_data['Case']), 'train', 'test')
        train = encoder_df_temp[encoder_df_temp['set'] == 'train']
        train = train.drop(['set'], axis=1)

        # train 데이터를 이어붙이기
        encoder_df = pd.concat([encoder_df, train])

        encoder_df = encoder_df.fillna(0)
        encoder_df = encoder_df.astype(int)
    
        # 결과 저장
        encoder_df.to_csv(concat_path + "\\" + noise_type[i] + '_BPIC15_1_' + 'target_ratio_40_concat.csv', index=True) # change (3)

import os
os.chmod('c:\\Users\\yeon1\\data-noise-research2\\concat', 0o700)
